+++ server/src/index.ts
@@ -14,8 +14,8 @@ const options: cors.CorsOptions = {
 
 const app: Application = express()
 app.use(cors(options))
-app.use(express.json())
-app.use(express.urlencoded({ extended: true, limit: 50000 }))
+app.use(express.json({ limit: '10mb' }))
+app.use(express.urlencoded({ limit: '10mb' }))
 
 app.use('/', [ingredientRoutes, recipeRoutes])